+++ models/beneficiary.model.js
@@ -158,14 +158,108 @@ module.exports = class Beneficiary {
             const rows = await database.query(
                 `SELECT Ben.*, LD.nombre AS discapacidad
                 FROM Beneficiarios Ben
-                INNER JOIN BeneficiarioDiscapacidades BD ON Ben.idBeneficiario = BD.idBeneficiario
-                INNER JOIN ListaDiscapacidades LD ON LD.idDiscapacidad = BD.idDiscapacidad
+                LEFT JOIN BeneficiarioDiscapacidades BD ON Ben.idBeneficiario = BD.idBeneficiario
+                LEFT JOIN ListaDiscapacidades LD ON LD.idDiscapacidad = BD.idDiscapacidad
+                WHERE estatus = 1
                 ORDER BY Ben.nombre`);
             console.log("ROWS:", rows);
             return rows;
         } catch (err) {
             console.error("Error al obtener beneficiarios:", err);
             throw err; 
+        }
+    }
+
+    // Model para BEN-003
+    static async update(beneficiaryId, nombre, apellidoPaterno, apellidoMaterno, fechaNacimiento, numeroEmergencia, nombreContactoEmergencia, relacionContactoEmergencia, descripcion, fechaIngreso, foto, estatus) {
+        try {
+            const query = `UPDATE Beneficiarios
+                           SET nombre = ?,
+                           apellidoPaterno = ?,
+                           apellidoMaterno = ?,
+                           fechaNacimiento = ?,
+                           numeroEmergencia = ?,
+                           nombreContactoEmergencia = ?,
+                           relacionContactoEmergencia = ?,
+                           descripcion = ?,
+                           fechaIngreso = ?,
+                           foto = ?,
+                           estatus  = ?
+                           WHERE idBeneficiario = ?`;
+
+                           const params = [nombre, apellidoPaterno, apellidoMaterno, fechaNacimiento, numeroEmergencia, nombreContactoEmergencia, relacionContactoEmergencia, descripcion, fechaIngreso, foto, estatus, beneficiaryId];
+                           const result = await database.execute(query, params);
+                           return result;
+        } catch (err) {
+            console.error(`Error al actualizar beneficiario con id ${beneficiaryId}:`, err);
+            throw err;
+        }
+    }
+
+    // Filter beneficiaries by disability (list ordered by name)
+    static async filter(body = {}){
+        const filters = body.filter;
+        try{
+            let query = `
+            SELECT Ben.*, LD.nombre AS discapacidad
+            FROM Beneficiarios Ben
+            LEFT JOIN BeneficiarioDiscapacidades BD ON Ben.idBeneficiario = BD.idBeneficiario
+            LEFT JOIN ListaDiscapacidades LD ON LD.idDiscapacidad = BD.idDiscapacidad
+            WHERE estatus = 1
+            `;
+            const  params =[];
+            if (filters["disability"] && filters["disability"].length > 0) {
+                query += ` AND LD.nombre IN (${filters["disability"].map(() => '?').join(', ')})`;
+                params.push(...filters["disability"]);
+            }
+
+            if (body.order){
+                query += ` ORDER BY Ben.nombre ${body.order}`;
+            }
+            const rows = await database.query(query, params);
+            return rows;
+        }
+        catch(error){
+            console.log("Error al obtener lista de beneficiarios", error);
+            throw error;
+        }
+    }
+
+    // Get categories on the disabilities for filtering
+    static async getCategories(){
+        try{
+            const discapacidad = await database.query(
+                `
+                SELECT DISTINCT nombre as discapacidad
+                FROM ListaDiscapacidades
+                `
+            );
+
+            return {
+                    discapacidad: discapacidad.map(e => e.discapacidad),
+                };
+        } catch(error){
+            console.error("Error al obtener datos de filtrado:", error);
+            throw error;
+        }
+    }
+
+    // --- Nuevo para BEN-007
+    static async searchByName(searchTerm) {
+        try {
+            const query = `
+                SELECT * FROM Beneficiarios
+                WHERE CONCAT(nombre, ' ', apellidoPaterno, ' ', apellidoMaterno) LIKE ?
+                AND estatus = 1
+            `;
+            const params = [`%${searchTerm}%`];
+
+            const rows = await database.query(query, params);
+
+            return rows;
+        } catch (err) {
+            console.error(`Error al buscar beneficiarios con termino "${searchTerm}":`, err);
+            throw err;
         }
     }
 }