+++ routes/workshops.routes.js
@@ -13,4 +13,11 @@ router.post('/modify/:idTaller', workshopsController.modifyWorkshops);
 
 router.put('/delete/:idTaller', workshopsController.deleteWorkshops);
 
+/* Get categories of workshops in horaEntrada*/
+router.get("/filter/data", workshopsController.getWorkshopsCategories);
+
+/* POST - Filters Workshops */
+router.post('/filter', workshopsController.viewWorkshopsFiltered);
+
+
 module.exports = router;